:white_check_mark: Move all tests to index.spec.tsx
