+++ static/app/components/searchQueryBuilder/index.spec.tsx
@@ -1,6 +1,7 @@
 import type {ComponentProps} from 'react';
 import {destroyAnnouncer} from '@react-aria/live-announcer';
 import {AutofixSetupFixture} from 'sentry-fixture/autofixSetupFixture';
+import {OrganizationFixture} from 'sentry-fixture/organization';
 
 import {
   act,
@@ -946,52 +947,6 @@ describe('SearchQueryBuilder', function () {
 
       // Should have a filter token "browser.name:foo"
       expect(screen.getByRole('row', {name: 'browser.name:foo'})).toBeInTheDocument();
-    });
-
-    it('displays ask seer button when searching free text', async function () {
-      const mockOnSearch = jest.fn();
-      MockApiClient.addMockResponse({
-        url: '/organizations/org-slug/seer/setup-check/',
-        body: AutofixSetupFixture({
-          setupAcknowledgement: {
-            orgHasAcknowledged: true,
-            userHasAcknowledged: true,
-          },
-        }),
-      });
-
-      render(
-        <SearchQueryBuilder {...defaultProps} enableAISearch onSearch={mockOnSearch} />,
-        {organization: {features: ['gen-ai-features', 'gen-ai-explore-traces']}}
-      );
-
-      await userEvent.click(getLastInput());
-      await userEvent.type(screen.getByRole('combobox'), 'some free text');
-
-      expect(screen.getByRole('option', {name: /Ask Seer/i})).toBeInTheDocument();
-    });
-
-    it('displays enable ai button when searching free text and user has not given consent', async function () {
-      const mockOnSearch = jest.fn();
-      MockApiClient.addMockResponse({
-        url: '/organizations/org-slug/seer/setup-check/',
-        body: AutofixSetupFixture({
-          setupAcknowledgement: {
-            orgHasAcknowledged: false,
-            userHasAcknowledged: false,
-          },
-        }),
-      });
-
-      render(
-        <SearchQueryBuilder {...defaultProps} enableAISearch onSearch={mockOnSearch} />,
-        {organization: {features: ['gen-ai-features', 'gen-ai-explore-traces']}}
-      );
-
-      await userEvent.click(getLastInput());
-      await userEvent.type(screen.getByRole('combobox'), 'some free text');
-
-      expect(screen.getByRole('option', {name: /Enable Gen AI/})).toBeInTheDocument();
     });
 
     it('can add parens by typing', async function () {
@@ -4475,6 +4430,143 @@ describe('SearchQueryBuilder', function () {
       expect(
         screen.getByRole('row', {name: 'span.description:"random value"'})
       ).toBeInTheDocument();
+    });
+  });
+
+  describe('ask seer', function () {
+    it('renders ask seer button when user has given consent', async () => {
+      MockApiClient.addMockResponse({
+        url: '/organizations/org-slug/seer/setup-check/',
+        body: AutofixSetupFixture({
+          setupAcknowledgement: {
+            orgHasAcknowledged: true,
+            userHasAcknowledged: true,
+          },
+        }),
+      });
+
+      render(<SearchQueryBuilder {...defaultProps} enableAISearch />, {
+        organization: {features: ['gen-ai-features', 'gen-ai-explore-traces']},
+      });
+
+      await userEvent.click(getLastInput());
+
+      const askSeer = await screen.findByRole('option', {name: /Ask Seer/});
+      expect(askSeer).toBeInTheDocument();
+    });
+
+    it('renders enable ai button when user has not given consent', async () => {
+      MockApiClient.addMockResponse({
+        url: '/organizations/org-slug/seer/setup-check/',
+        body: AutofixSetupFixture({
+          setupAcknowledgement: {
+            orgHasAcknowledged: false,
+            userHasAcknowledged: false,
+          },
+        }),
+      });
+
+      render(<SearchQueryBuilder {...defaultProps} enableAISearch />, {
+        organization: {features: ['gen-ai-features', 'gen-ai-explore-traces']},
+      });
+
+      await userEvent.click(getLastInput());
+
+      const enableAi = await screen.findByText(/Enable Gen AI/);
+      expect(enableAi).toBeInTheDocument();
+    });
+
+    describe('user clicks on enable gen ai button', () => {
+      it('calls promptsUpdate', async () => {
+        const organization = OrganizationFixture();
+        MockApiClient.addMockResponse({
+          url: '/organizations/org-slug/seer/setup-check/',
+          body: AutofixSetupFixture({
+            setupAcknowledgement: {
+              orgHasAcknowledged: false,
+              userHasAcknowledged: false,
+            },
+          }),
+        });
+
+        const promptsUpdateMock = MockApiClient.addMockResponse({
+          url: `/organizations/${organization.slug}/prompts-activity/`,
+          method: 'PUT',
+        });
+
+        render(<SearchQueryBuilder {...defaultProps} enableAISearch />, {
+          organization: {features: ['gen-ai-features', 'gen-ai-explore-traces']},
+        });
+
+        await userEvent.click(getLastInput());
+
+        const enableAi = await screen.findByText(/Enable Gen AI/);
+        expect(enableAi).toBeInTheDocument();
+
+        await userEvent.click(enableAi);
+
+        await waitFor(() => {
+          expect(promptsUpdateMock).toHaveBeenCalledWith(
+            expect.any(String),
+            expect.objectContaining({
+              data: {
+                feature: 'seer_autofix_setup_acknowledged',
+                organization_id: organization.id,
+                project_id: undefined,
+                status: 'dismissed',
+              },
+            })
+          );
+        });
+      });
+    });
+
+    describe('free text', function () {
+      it('displays ask seer button when searching free text', async function () {
+        const mockOnSearch = jest.fn();
+        MockApiClient.addMockResponse({
+          url: '/organizations/org-slug/seer/setup-check/',
+          body: AutofixSetupFixture({
+            setupAcknowledgement: {
+              orgHasAcknowledged: true,
+              userHasAcknowledged: true,
+            },
+          }),
+        });
+
+        render(
+          <SearchQueryBuilder {...defaultProps} enableAISearch onSearch={mockOnSearch} />,
+          {organization: {features: ['gen-ai-features', 'gen-ai-explore-traces']}}
+        );
+
+        await userEvent.click(getLastInput());
+        await userEvent.type(screen.getByRole('combobox'), 'some free text');
+
+        expect(screen.getByRole('option', {name: /Ask Seer/i})).toBeInTheDocument();
+      });
+    });
+
+    it('displays enable ai button when searching free text and user has not given consent', async function () {
+      const mockOnSearch = jest.fn();
+      MockApiClient.addMockResponse({
+        url: '/organizations/org-slug/seer/setup-check/',
+        body: AutofixSetupFixture({
+          setupAcknowledgement: {
+            orgHasAcknowledged: false,
+            userHasAcknowledged: false,
+          },
+        }),
+      });
+
+      render(
+        <SearchQueryBuilder {...defaultProps} enableAISearch onSearch={mockOnSearch} />,
+        {organization: {features: ['gen-ai-features', 'gen-ai-explore-traces']}}
+      );
+
+      await userEvent.click(getLastInput());
+      await userEvent.type(screen.getByRole('combobox'), 'some free text');
+
+      expect(screen.getByRole('option', {name: /Enable Gen AI/})).toBeInTheDocument();
     });
   });
 });
